+++ lib/features/authentication/screens/login/widgets/login_header.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+import '../../../../../utils/constants/image_strings.dart';
+import '../../../../../utils/constants/sizes.dart';
+import '../../../../../utils/constants/text_strings.dart';
+
+class TLoginHeader extends StatelessWidget {
+  const TLoginHeader({
+    super.key,
+    required this.dark,
+  });
+
+  final bool dark;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Padding(
+          padding: const EdgeInsets.symmetric(vertical:TSizes.spaceBtwItems), // Default padding, you can adjust this
+          child: Image(
+            height: 80,
+            image: AssetImage(dark ? TImages.lightAppLogo : TImages.darkAppLogo),
+          ),
+        ),
+        Text(TTexts.loginTitle, style: Theme.of(context).textTheme.headlineMedium),
+        Text(TTexts.loginSubTitle, style: Theme.of(context).textTheme.bodyMedium),
+      ],
+    );
+  }
+}
